move config module to the util to avoid name conflicts
