+++ test.py
@@ -6,7 +6,7 @@ import numpy as np
 import scipy.io
 import scipy.ndimage
 
-from config import load_config
+from util.config import load_config
 from dataset.factory import create as create_dataset
 from dataset.pose_dataset import Batch
 from nnet.predict import setup_pose_prediction, extract_cnn_output, argmax_pose_predict
