+++ train.py
@@ -4,7 +4,7 @@ import threading
 import tensorflow as tf
 import tensorflow.contrib.slim as slim
 
-from config import load_config
+from util.config import load_config
 from dataset.factory import create as create_dataset
 from nnet.net_factory import pose_net
 from nnet.pose_net import get_batch_spec
