+++ util/config.py
@@ -5,10 +5,10 @@ import logging
 import yaml
 from easydict import EasyDict as edict
 
-import default_config
+import util.default_config
 
 
-cfg = default_config.cfg
+cfg = util.default_config.cfg
 
 
 def _merge_a_into_b(a, b):
